vernieuwing-inzage-overzicht/backend/pyproject.toml
[tool.poetry]
name = "Local development vernieuwing-inzage"
version = "0.1.0"
description = ""
authors = ["ictu devops"]

[tool.poetry.dependencies]
python = "^3.11"
SQLAlchemy = "^2.0.7"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
pandas = "^1.5.1"
SQLAlchemy-Utils = "^0.40.0"
secure = "^0.3.0"
PyJWT = "^2.6.0"
wheel = "^0.38.1"
pytest = "^7.4.0"
coverage = "^7.2.7"
requests = "^2.31.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
httpx = "^0.26.0"
alembic = "^1.8.1"
pydantic-settings = "^2.2.1"
pyotp = "^2.9.0"
asyncpg = "^0.29.0"

[tool.poetry.dev-dependencies]
black = {extras = ["jupyter"], version = "^22.10.0"}
flake8 = "^5.0.4"
ipykernel = "^6.16.2"
ruff = "^0.1.11"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
# Enable flake8-bugbear (`B`) rules, in addition to the defaults.
select = ['E', 'F', 'W', 'A', 'PLC', 'PLE', 'PLW', 'I']

# Avoid enforcing line-length violations (`E501`)
ignore = ["E501", "D104", "D100", "D107", "D103"]

# Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

# Ignore `F401` (import violations) in all `__init__.py` files, and in select subdirectories.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "I001"]

[tool.ruff.format]
# Use single quotes for non-triple-quoted strings.
quote-style = "double"

[tool.coverage.run]
omit = [
     "**/frontend/**/*",
     "**/__init__.py",
     "**/models/**/*",
     "**/schemas/**/*",
    ]